INITIAL

+++ src/main/java/com/stackroute/service/TrackServiceImpl.java
@@ -1,6 +1,8 @@
 package com.stackroute.service;
 
 import com.stackroute.domain.Track;
+import com.stackroute.exception.TrackAlreadyExists;
+import com.stackroute.exception.TrackNotFound;
 import com.stackroute.repository.TrackRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -20,9 +22,13 @@ public class TrackServiceImpl implements TrackService {
     }
 
     @Override
-    public Track saveTrack(Track track) {
-       Track savedTrack = trackRepository.save(track);
-       return savedTrack;
+    public Track saveTrack(Track track) throws TrackAlreadyExists {
+        if (trackRepository.existsById(track.getTrackId())) {
+            throw new TrackAlreadyExists("TrackAlreadyExists");
+        } else {
+            Track savedTrack = trackRepository.save(track);
+            return savedTrack;
+        }
     }
 
     @Override
@@ -32,29 +38,42 @@ public class TrackServiceImpl implements TrackService {
     }
 
     @Override
-    public Optional<Track> getTrackById(int id) {
-
-        Optional<Track> getTrack = trackRepository.findById(id);
-        return  getTrack;
+    public Optional<Track> getTrackById(int id) throws TrackNotFound {
+        if (trackRepository.existsById(id)) {
+            Optional<Track> getTrack = trackRepository.findById(id);
+            return getTrack;
+        } else {
+            throw new TrackNotFound("Track Not Found");
+        }
     }
 
     @Override
-    public Track updateComments(Track track) {
-        trackRepository.delete(track);
-      Track updatedTrack  =saveTrack(track);
-     return updatedTrack;
+    public Track updateComments(Track track) throws TrackAlreadyExists {
+
+        if (trackRepository.existsById(track.getTrackId())) {
+            throw new TrackAlreadyExists("TrackAlreadyExists");
+        } else {
+            trackRepository.delete(track);
+            Track savedTrack = trackRepository.save(track);
+            return savedTrack;
+        }
+
 
     }
 
     @Override
-    public  void deleteTrackById(int id) {
-
-         trackRepository.deleteById(id);
+    public void deleteTrackById(int id) throws TrackNotFound {
+        if (trackRepository.existsById(id)) {
+            trackRepository.deleteById(id);
+        } else {
+            throw new TrackNotFound("Track Not Found");
+        }
     }
 
     @Override
     public List<Track> findByTrackName(String trackName) {
         return trackRepository.findByTrackName(trackName);
     }
+
 
 }
